Conflicts:
	bcl2fastq/Snakefile

+++ variant-calling/gatk/tests.sh
@@ -6,6 +6,9 @@
 set -euo pipefail
 
 MYNAME=$(basename $(readlink -f $0))
+#PIPELINE=$(basename $(dirname $MYNAME))
+PIPELINE=$(basename $(dirname $(readlink -f $0)))
+DOWNSTREAM_OUTDIR_PY=$(readlink -f $(dirname $MYNAME)/../../tools/downstream_outdir.py)
 
 toaddr() {
     if [ $(whoami) == 'userrig' ]; then
@@ -48,6 +51,8 @@ WGS_FQ1=$RPD_ROOT/testing/data/illumina-platinum-NA12878/ERR091571_1.fastq.gz
 WGS_FQ2=$RPD_ROOT/testing/data/illumina-platinum-NA12878/ERR091571_2.fastq.gz
 DUMMY_BED=$RPD_ROOT/testing/data/illumina-platinum-NA12878/human_g1k_v37_decoy_chr21.bed
 TRUSEQ_BED=$RPD_ROOT/testing/data/illumina-platinum-NA12878/exome/truseq-exome-targeted-regions-manifest-v1-2.nochr.bed
+DREAM_WGS_DIR=$RPD_ROOT/testing/data/somatic/icgc-tcga-dream-somatic/synthetic.challenge.set3/
+INJ_BAM=$DREAM_WGS_DIR/normal.bam
 
 rootdir=$(readlink -f $(dirname $0))
 cd $rootdir
@@ -56,7 +61,6 @@ commit=$(git describe --always --dirty)
 test -e Snakefile || exit 1
 
 
-test_outdir_base=$RPD_ROOT/testing/output/${pipeline}/${pipeline}-commit-${commit}
 log=$(mktemp)
 COMPLETE_MSG="*** All tests completed ***"
 echo "Logging to $log"
@@ -79,12 +83,12 @@ wgs_cmd_base="$WRAPPER -1 $WGS_FQ1 -2 $WGS_FQ2 -s NA12878-WGS -t WGS --name 'tes
 SKIP_DAG=1
 if [ $SKIP_DAG -eq 0 ]; then
     echo "DAG: WES" | tee -a $log
-    odir=$(mktemp -d ${test_outdir_base}-wes.XXXXXXXXXX) && rmdir $odir
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
     eval $wes_cmd_base -o $odir -v --no-run >> $log 2>&1
     pushd $odir >> $log
     type=pdf;
     dag=example-dag.$type
-    sed -i -e 's,num_chroms: .*,num_chroms: 1,' conf.yaml
+    # FIXME simplify DAG by having only one cluster sed -i -e 's,num_chroms: .*,num_chroms: 1,' conf.yaml
     EXTRA_SNAKEMAKE_ARGS="--dag" bash run.sh; cat logs/snakemake.log | dot -T$type > $dag
     cp $dag $rootdir
     rm -rf $odir
@@ -96,16 +100,46 @@ fi
 #
 if [ $skip_dry_runs -ne 1 ]; then
     echo "Dryrun: targeted" | tee -a $log
-    odir=$(mktemp -d ${test_outdir_base}-targeted.XXXXXXXXXX) && rmdir $odir
-    eval $targeted_cmd_base -o $odir -v --no-run >> $log 2>&1
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
+    # also testing --extra-conf
+    eval $targeted_cmd_base -o $odir -v --extra-conf extrakey:extravalue --no-run >> $log 2>&1
     pushd $odir >> $log
+    grep -q extrakey conf.yaml && echo "extra-conf works" >> $log 2>&1 || exit 1
     EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
     popd >> $log
     rm -rf $odir
     
     echo "Dryrun: WES" | tee -a $log
-    odir=$(mktemp -d ${test_outdir_base}-wes.XXXXXXXXXX) && rmdir $odir
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
     eval $wes_cmd_base -o $odir -v --no-run >> $log 2>&1
+    pushd $odir >> $log
+    EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
+    popd >> $log
+    rm -rf $odir
+
+    echo "Dryrun: WES bam only" | tee -a $log
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
+    eval $wes_cmd_base -o $odir --bam-only -v --no-run >> $log 2>&1
+    pushd $odir >> $log
+    EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
+    popd >> $log
+    rm -rf $odir
+
+    echo "Dryrun: WES injected raw" | tee -a $log
+
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
+    _wes_cmd_base="$WRAPPER -s NA12878-WES -t WES -l $TRUSEQ_BED --name 'test:WES'";# --bam-only"
+    eval $_wes_cmd_base -o $odir --raw-bam $INJ_BAM -v --no-run >> $log 2>&1
+
+    pushd $odir >> $log
+    EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
+    popd >> $log
+    rm -rf $odir
+
+    echo "Dryrun: WES injected post processed" | tee -a $log
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
+    eval $_wes_cmd_base -o $odir --proc-bam $INJ_BAM -v --no-run >> $log 2>&1
+
     pushd $odir >> $log
     EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
     popd >> $log
@@ -113,7 +147,7 @@ if [ $skip_dry_runs -ne 1 ]; then
 
     # FIXME run wrapper should be tested somewhere else
     echo "Dryrun: WES through run wrapper" | tee -a $log
-    odir=$(mktemp -d ${test_outdir_base}-wes-run.XXXXXXXXXX) && rmdir $odir
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
     eval $wes_run_cmd_base -o $odir -v --no-run >> $log 2>&1
     pushd $odir >> $log
     EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
@@ -121,7 +155,7 @@ if [ $skip_dry_runs -ne 1 ]; then
     rm -rf $odir
 
     echo "Dryrun: WGS" | tee -a $log
-    odir=$(mktemp -d ${test_outdir_base}-wgs.XXXXXXXXXX) && rmdir $odir
+    odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
     eval $wgs_cmd_base -o $odir -v --no-run >> $log 2>&1
     pushd $odir >> $log
     EXTRA_SNAKEMAKE_ARGS="--dryrun" bash run.sh >> $log 2>&1
@@ -138,7 +172,7 @@ fi
 if [ $skip_real_runs -ne 1 ]; then
     if [ $SKIP_REAL_TARGETED -eq 0 ]; then
         echo "Realrun: targeted" | tee -a $log
-        odir=$(mktemp -d ${test_outdir_base}-targeted.XXXXXXXXXX) && rmdir $odir
+	odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
         eval $targeted_cmd_base -o $odir -v >> $log 2>&1
         # magically works even if line just contains id as in the case of pbspro
         jid=$(tail -n 1 $odir/logs/submission.log  | cut -f 3 -d ' ')
@@ -149,7 +183,7 @@ if [ $skip_real_runs -ne 1 ]; then
     
     if [ $SKIP_REAL_WES -eq 0 ]; then
         echo "Realrun: WES" | tee -a $log
-        odir=$(mktemp -d ${test_outdir_base}-wes.XXXXXXXXXX) && rmdir $odir
+	odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
         eval $wes_cmd_base -o $odir -v >> $log 2>&1
         # magically works even if line just contains id as in the case of pbspro
         jid=$(tail -n 1 $odir/logs/submission.log  | cut -f 3 -d ' ')
@@ -161,7 +195,7 @@ if [ $skip_real_runs -ne 1 ]; then
 
     if [ $SKIP_REAL_WGS -eq 0 ]; then
         echo "Realrun: WGS" | tee -a $log
-        odir=$(mktemp -d ${test_outdir_base}-wgs.XXXXXXXXXX) && rmdir $odir
+	odir=$($DOWNSTREAM_OUTDIR_PY -r $(whoami) -p $PIPELINE)
         eval $wgs_cmd_base -o $odir -v >> $log 2>&1
         # magically works even if line just contains id as in the case of pbspro
         jid=$(tail -n 1 $odir/logs/submission.log  | cut -f 3 -d ' ')
